core: split core/arch/arm/include/kernel/thread.h
Splits core/arch/arm/include/kernel/thread.h into one generic and one
architecture specific file.

Acked-by: Marouene Boubakri <marouene.boubakri@nxp.com>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
